+++ tests/testthat/test-check_micromamba_version.R
@@ -48,8 +48,8 @@ testthat::test_that("is_umamba_version_available accepts newer versions", {
   # TODO: @luciorq - Update this to the newer version of micromamba upon
   # + update and ensure the test reflects that
   testthat::local_mocked_bindings(
-    get_micromamba_version = function(...) "2.6.2"
+    get_micromamba_version = function(...) "2.8.1"
   )
-  # This should now return TRUE since check_micromamba_version uses >= by default
+  # This should return TRUE since check_micromamba_version uses >= by default
   testthat::expect_true(is_umamba_version_available("/mock/path"))
 })